openssl: Enable universal build on 10.4
WIP, not for merging.

This commit hardcodes the universal architectures to experiment with
building the OpenSSL formula for all the 4 architectures suppirted on
Tiger.

In order to actually build the universal library, the formula must be
built using the flags `--universal --without-test`, because at least
one set of tests cannot be run (if built with testing enabled on an
`x86_64` machine, the only failing test should be `ppc64`).